+++ python_bindings/src/PyIROperator.cpp
@@ -5,77 +5,6 @@ namespace PythonBindings {
 
 namespace {
 
-Expr select0(Expr condition, Expr true_value, Expr false_value) {
-    return select(condition, true_value, false_value);
-}
-
-Expr select1(Expr c1, Expr v1,
-                Expr c2, Expr v2,
-                Expr default_val) {
-    return select(c1, v1,
-                     c2, v2, default_val);
-}
-Expr select2(Expr c1, Expr v1,
-                Expr c2, Expr v2,
-                Expr c3, Expr v3,
-                Expr default_val) {
-    return select(c1, v1,
-                     c2, v2,
-                     c3, v3, default_val);
-}
-Expr select3(Expr c1, Expr v1,
-                Expr c2, Expr v2,
-                Expr c3, Expr v3,
-                Expr c4, Expr v4,
-                Expr default_val) {
-    return select(c1, v1,
-                     c2, v2,
-                     c3, v3,
-                     c4, v4, default_val);
-}
-Expr select4(Expr c1, Expr v1,
-                Expr c2, Expr v2,
-                Expr c3, Expr v3,
-                Expr c4, Expr v4,
-                Expr c5, Expr v5,
-                Expr default_val) {
-    return select(c1, v1,
-                     c2, v2,
-                     c3, v3,
-                     c4, v4,
-                     c5, v5, default_val);
-}
-Expr select5(Expr c1, Expr v1,
-                Expr c2, Expr v2,
-                Expr c3, Expr v3,
-                Expr c4, Expr v4,
-                Expr c5, Expr v5,
-                Expr c6, Expr v6,
-                Expr default_val) {
-    return select(c1, v1,
-                     c2, v2,
-                     c3, v3,
-                     c4, v4,
-                     c5, v5,
-                     c6, v6, default_val);
-}
-Expr select6(Expr c1, Expr v1,
-                Expr c2, Expr v2,
-                Expr c3, Expr v3,
-                Expr c4, Expr v4,
-                Expr c5, Expr v5,
-                Expr c6, Expr v6,
-                Expr c7, Expr v7,
-                Expr default_val) {
-    return select(c1, v1,
-                     c2, v2,
-                     c3, v3,
-                     c4, v4,
-                     c5, v5,
-                     c6, v6,
-                     c7, v7, default_val);
-}
-
 // TODO: clever template usage could generalize this to list-of-types-to-try.
 std::vector<Expr> args_to_vector_for_print(const py::args &args, size_t start_offset = 0) {
     if (args.size() < start_offset) {
@@ -99,24 +28,50 @@ std::vector<Expr> args_to_vector_for_print(const py::args &args, size_t start_of
 }  // namespace
 
 void define_operators(py::module &m) {
-    m.def("max", (Expr (*)(Expr, Expr)) &max);
-    m.def("max", (Expr (*)(Expr, int)) &max);
-    m.def("max", (Expr (*)(int, Expr)) &max);
-    m.def("min", (Expr (*)(Expr, Expr)) &min);
-    m.def("min", (Expr (*)(Expr, int)) &min);
-    m.def("min", (Expr (*)(int, Expr)) &min);
+    m.def("max", [](py::args args) -> Expr {
+        if (args.size() < 2) {
+            throw py::value_error("max() must have at least 2 arguments");
+        }
+        int pos = (int) args.size() - 1;
+        Expr value = args[pos--].cast<Expr>();
+        while (pos >= 0) {
+            value = max(args[pos--].cast<Expr>(), value);
+        }
+        return value;
+    });
+
+    m.def("min", [](py::args args) -> Expr {
+        if (args.size() < 2) {
+            throw py::value_error("min() must have at least 2 arguments");
+        }
+        int pos = (int) args.size() - 1;
+        Expr value = args[pos--].cast<Expr>();
+        while (pos >= 0) {
+            value = min(args[pos--].cast<Expr>(), value);
+        }
+        return value;
+    });
+
     m.def("clamp", &clamp);
     m.def("abs", &abs);
     m.def("absd", &absd);
 
-    // TODO: improve this to use a single overload and py::args
-    m.def("select", &select0);
-    m.def("select", &select1);
-    m.def("select", &select2);
-    m.def("select", &select3);
-    m.def("select", &select4);
-    m.def("select", &select5);
-    m.def("select", &select6);
+    m.def("select", [](py::args args) -> Expr {
+        if (args.size() < 3) {
+            throw py::value_error("select() must have at least 3 arguments");
+        }
+        if ((args.size() % 2) == 0) {
+            throw py::value_error("select() must have an odd number of arguments");
+        }
+        int pos = (int) args.size() - 1;
+        Expr false_value = args[pos--].cast<Expr>();
+        while (pos > 0) {
+            Expr true_value = args[pos--].cast<Expr>();
+            Expr condition = args[pos--].cast<Expr>();
+            false_value = select(condition, true_value, false_value);
+        }
+        return false_value;
+    });
 
     m.def("sin", &sin);
     m.def("asin", &asin);
@@ -151,29 +106,36 @@ void define_operators(py::module &m) {
     m.def("is_nan", &is_nan);
     m.def("reinterpret", (Expr (*)(Type, Expr)) &reinterpret);
     m.def("cast", (Expr (*)(Type, Expr)) &cast);
-
-    // variadic args need a little extra love.
     m.def("print", [](py::args args) -> Expr {
         return print(args_to_vector_for_print(args));
     });
-
-    // variadic args need a little extra love.
-    m.def("print_when", [](py::args args) -> Expr {
-        Expr cond = args[0].cast<Expr>();
-        return print_when(cond, args_to_vector_for_print(args, 1));
-    });
-
+    m.def("print_when", [](Expr condition, py::args args) -> Expr {
+        return print_when(condition, args_to_vector_for_print(args));
+    }, py::arg("condition"));
+    m.def("require", [](Expr condition, Expr value, py::args args) -> Expr {
+        auto v = args_to_vector<Expr>(args);
+        v.insert(v.begin(), value);
+        return require(condition, v);
+    }, py::arg("condition"), py::arg("value"));
     m.def("lerp", &lerp);
     m.def("popcount", &popcount);
     m.def("count_leading_zeros", &count_leading_zeros);
     m.def("count_trailing_zeros", &count_trailing_zeros);
-    m.def("random_float", (Expr (*)(Expr)) &random_float);
+    m.def("div_round_to_zero", &div_round_to_zero);
+    m.def("mod_round_to_zero", &mod_round_to_zero);
     m.def("random_float", (Expr (*)()) &random_float);
-    m.def("random_int", (Expr (*)(Expr)) &random_int);
+    m.def("random_uint", (Expr (*)()) &random_uint);
     m.def("random_int", (Expr (*)()) &random_int);
+    m.def("random_float", (Expr (*)(Expr)) &random_float, py::arg("seed"));
+    m.def("random_uint", (Expr (*)(Expr)) &random_uint, py::arg("seed"));
+    m.def("random_int", (Expr (*)(Expr)) &random_int, py::arg("seed"));
     m.def("undef", (Expr (*)(Type)) &undef);
-
+    m.def("memoize_tag", [](Expr result, py::args cache_key_values) -> Expr {
+        return Internal::memoize_tag_helper(result, args_to_vector<Expr>(cache_key_values));
+    }, py::arg("result"));
     m.def("likely", &likely);
+    m.def("likely_if_innermost", &likely_if_innermost);
+    m.def("saturating_cast", (Expr (*)(Type, Expr))&saturating_cast);
 }
 
 }  // namespace PythonBindings
